Comment and clean circuit_gen, move to new file qiskit_extra.py
