+++ nchoosek/solver/__init__.py
@@ -1,1 +1,2 @@
 from nchoosek.solver.common import *
+from nchoosek.solver.qiskit_extra import circuit_gen
